ref: Zerologlevel gets own file
