+++ pkg/config/config.go
@@ -13,25 +13,6 @@ import (
 	"k8s.io/client-go/util/homedir"
 )
 
-type ZeroLogLevel int8
-
-func (l ZeroLogLevel) String() string {
-	return zerolog.Level(l).String()
-}
-
-func (l *ZeroLogLevel) Set(s string) error {
-	level, err := zerolog.ParseLevel(s)
-	if err != nil {
-		return err
-	}
-	*l = ZeroLogLevel(level)
-	return nil
-}
-
-func (l ZeroLogLevel) Type() string {
-	return "string"
-}
-
 type Config struct {
 	AdditionalKinds []string
 	Cluster         bool
